// src/types/post-block-update-types.ts
export enum PostBlockActionType {
  PdfGeneratePresignedUploadUrl = "PDF_GENERATE_PRESIGNED_UPLOAD_URL",
  GeneratePresignedUploadUrl = "GENERATE_PRESIGNED_UPLOAD_URL",
  StopQueryGeneration = "STOP_QUERY_GENERATION",
  GenerateDescription = "GENERATE_DESCRIPTION",
  StopCodeGeneration = "STOP_CODE_GENERATION",
  EditResultVariable = "EDIT_RESULT_VARIABLE",
  AnswerPdfQuestion = "ANSWER_PDF_QUESTION",
  PaginateDataframe = "PAGINATE_DATAFRAME",
  RunPythonBlock = "RUN_PYTHON_BLOCK",
  RunChartBlock = "RUN_CHART_BLOCK",
  RunTableBlock = "RUN_TABLE_BLOCK",
  UnverifyQuery = "UNVERIFY_QUERY",
  FixPython = "AUTO_FIX_PYTHON",
  RunCsvBlock = "RUN_CSV_BLOCK",
  RunSqlBlock = "RUN_SQL_BLOCK",
  UpdateBlock = "UPDATE_BLOCK",
  WritePython = "WRITE_PYTHON",
  VerifyQuery = "VERIFY_QUERY",
  DownloadCsv = "DOWNLOAD_CSV",
  DownloadSql = "DOWNLOAD_SQL",
  FixSql = "AUTO_FIX_SQL",
  IndexPDF = "INDEX_PDF",
  WriteSql = "WRITE_SQL",
}
// ...
export enum ResponseProjectActionType {
  // BotConversationMessage:
  CreateBotConversationMessage = "CREATE_BOT_CONVERSATION_MESSAGE",
  DeleteBotConversationMessage = "DELETE_BOT_CONVERSATION_MESSAGE",
  UpdateBotConversationMessage = "UPDATE_BOT_CONVERSATION_MESSAGE",
  UpdateBotConversation = "UPDATE_BOT_CONVERSATION",
  CreatePlanStep = "CREATE_PLAN_STEP",
  UpdatePlanStep = "UPDATE_PLAN_STEP",
  UpdatePlan = "UPDATE_PLAN",
  CreatePlan = "CREATE_PLAN",
  // Project:
  UpdateProject = "UPDATE_PROJECT",
  // Block:
  CreateBlock = "CREATE_BLOCK",
  UpdateBlock = "UPDATE_BLOCK",
  DeleteBlock = "DELETE_BLOCK",
}
// ...
export type CreatePlanStep = {
  action_type: ResponseProjectActionType.CreatePlanStep;
  action_payload: {
    plan_step: PlanStep;
  };
};

export type CreatePlan = {
  action_type: ResponseProjectActionType.CreatePlan;
  action_payload: {
    plan: Plan;
  };
};

export type UpdatePlanStep = {
  action_type: ResponseProjectActionType.UpdatePlanStep;
  action_payload: {
    value: string | Record<string, unknown> | unknown[] | number | boolean;
    plan_step_id: number;
    key: string[];
  };
};

export type UpdatePlan = {
  action_type: ResponseProjectActionType.UpdatePlan;
  action_payload: {
    value: string | Record<string, unknown> | unknown[] | number | boolean;
    plan_step_id: number;
    key: string[];
  };
};

export type UpdateBotConversationAction = {
  action_type: ResponseProjectActionType.UpdateBotConversation;
  action_payload: {
    value: boolean;
    key: string[];
    id: number;
  };
};

export type UpdateActionValue<T = unknown> =
  | { is_incremental_string_change?: false | null; value: T }
  | {
      is_incremental_string_change: true;
      value: string;
    };
// ...
export type UpdateBlockResponseAction<T = unknown> = {
  action_type: ResponseProjectActionType.UpdateBlock;
  action_payload: {
    block_uuid: NotebookBlockUuid;
    backend_overrides?: true;
    stream_uuid?: string;
    key: string[];
  } & UpdateActionValue<T>;
  timestamp: string;
};

export type PatchProjectResponseAction =
  // BotConversation:
  | CreateBotConversationMessageResponseAction
  | DeleteBotConversationMessageResponseAction
  | UpdateBotConversationMessageResponseAction
  | UpdateBotConversationAction
  | CreatePlanStep
  | UpdatePlanStep
  | CreatePlan
  | UpdatePlan
  // Project:
  | UpdateProjectResponseAction
  // Block:
  | CreateBlockResponseAction
  | DeleteBlockResponseAction
  | UpdateBlockResponseAction;

export type BasePostBlockActionOutput = {
  notebook_updates?: PatchProjectResponseAction[];
  error?: string | null;
  sources?: unknown[];
  answer?: string;
};
export type PostBlockResponse<T> = {
  action_output: BasePostBlockActionOutput & T;
  timestamp: ISODateString;
};